+++ src/provenance/MixedStorageProvenanceGraphManager.ts
@@ -1,17 +1,13 @@
 /**
  * Created by sam on 12.02.2015.
  */
-import {isFunction, constant, argList, mixin, search, hash, resolveIn} from '../index';
-import {get as getData, remove as removeData, upload, list as listData} from '../data';
-import * as graph from '../graph';
-import {IDType, SelectOperation, defaultSelectionType, resolve as resolveIDType} from '../idtype';
-import {Range, list as rlist, Range1D, all} from '../range';
-import {isDataType, IDataType, IDataDescription, DataTypeBase} from '../datatype';
-import {list as listPlugins, load as loadPlugin} from '../plugin';
-import * as session from '../session';
+import {IDataDescription} from '../datatype';
+import ProvenanceGraph, {IProvenanceGraphManager} from './ProvenanceGraph';
+import LocalStorageProvenanceGraphManager from './LocalStorageProvenanceGraphManager';
+import RemoteStorageProvenanceGraphManager from './RemoteStorageProvenanceGraphManager';
+import GraphBase from '../graph/GraphBase';
 
-
-export class MixedStorageProvenanceGraphManager implements IProvenanceGraphManager {
+export default class MixedStorageProvenanceGraphManager implements IProvenanceGraphManager {
   private remote:RemoteStorageProvenanceGraphManager;
   private local:LocalStorageProvenanceGraphManager;
 
@@ -48,7 +44,7 @@ export class MixedStorageProvenanceGraphManager implements IProvenanceGraphManag
     }
   }
 
-  getGraph(desc:IDataDescription):Promise<graph.GraphBase> {
+  getGraph(desc:IDataDescription):Promise<GraphBase> {
     if ((<any>desc).local) {
       return this.local.getGraph(desc);
     } else {
